feat: 오픈 채팅 2차 구현
