+++ src/app/allchats/page.tsx
@@ -1,6 +1,5 @@
-'use client';
+// 'use client';
 
-import React, { useEffect, useState } from 'react';
 import ChatPage from '@/components/chats/ChatPage';
 
 const AllChats = () => {
