+++ src/app/mychats/[userId].tsx
@@ -1,4 +1,4 @@
-'use client';
+// 'use client';
 
 import React, { useState } from 'react';
 import ChatPage from '@/components/chats/ChatPage';
