+++ src/components/chats/interfaces.ts
@@ -1,11 +1,8 @@
-import { instance } from '@/lib/api';
-import axios, { AxiosResponse } from 'axios';
-
 export interface Chat {
     id?: string;
     name?: string;
     isPrivate?: boolean;
-    users?: string[];
+    users?: User[];
     latestMessage?: Message | null;
 
     updatedAt?: Date;
@@ -25,13 +22,5 @@ export interface User {
     name: string;
     picture: string;
     chats: string[]; // chat id만 속합니다.
+    username: string;
 }
-
-// 채팅창에 보여줄 정보
-export interface ChatItem {
-    name?: string;
-    users?: User[];
-    latestMessage?: Message | null;
-}
-
-// 모든 채팅 정보 가져오는 함수
