+++ src/hooks/chatsList/useFormatCreatedAt.ts
@@ -1,0 +1,9 @@
+export const formatCreatedAt = (createdAt: Date) => {
+    const date = new Date(createdAt);
+    const hours = date.getHours();
+    const minutes = date.getMinutes();
+    const amOrPm = hours >= 12 ? '오후' : '오전';
+    const formattedHours = hours % 12 || 12;
+    const formattedMinutes = minutes < 10 ? `0${minutes}` : minutes;
+    return `${amOrPm} ${formattedHours}시 ${formattedMinutes}분`;
+};
